+++ CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.37.0](https://github.com/searchspring/snap/compare/v0.36.0...v0.37.0) (2022-09-26)
+
+
+### Features
+
+* **spell-correct:** initial commit of integrated spell correct feature (missing some testing) ([b4b265f](https://github.com/searchspring/snap/commit/b4b265f31728ca90bc26048b2ba8d458b5a85660))
+
+
+
+
+
+# [0.36.0](https://github.com/searchspring/snap/compare/v0.35.0...v0.36.0) (2022-09-19)
+
+
+### Bug Fixes
+
+* **preact-create-autocomplete:** fixing reversion bug that broke initializeFromUrl setting ([fda9884](https://github.com/searchspring/snap/commit/fda98842ff8b3ccde9318f26d4623abf77a0da49))
+* **recommendationcontroller.ts:** recommendations personalization for other siteIds ([c6be636](https://github.com/searchspring/snap/commit/c6be6367dc2e76b9fddfa86bda8040f8eb72e5e1))
+
+
+### Features
+
+* **searchrequest.ts:** adding the ability to disableInlineBanners ([5868195](https://github.com/searchspring/snap/commit/586819543455f384af0fb32b3a8765cf8ba2355e))
+
+
+
+
+
+# [0.35.0](https://github.com/searchspring/snap/compare/v0.34.7...v0.35.0) (2022-09-01)
+
+
+### Features
+
+* **getcontext.ts:** grab and return the siteId from the script context or src ([4830560](https://github.com/searchspring/snap/commit/4830560d1522b035d62ab5d4e3534e0371b30ac9))
+* **preact-components:** adding new hook 'useConstructor' for easy constructor like behavior ([a249f6a](https://github.com/searchspring/snap/commit/a249f6abb6c0ebca730563af05d9c344e1ae1071))
+* **searchstore:** add landingPage to the searchstore ([a7d2c4f](https://github.com/searchspring/snap/commit/a7d2c4fe2740168e436ed865e1a8b85120154691))
+
+
+
+
+
 ## [0.34.7](https://github.com/searchspring/snap/compare/v0.34.6...v0.34.7) (2022-08-19)
 
 **Note:** Version bump only for package @searchspring/snap